requirements.txt
pandas
numpy
yfinance
scikit-learn
torch
torchvision
matplotlib
seaborn
plotly
umap-learn
hdbscan
streamlit